+++ src/main/java/roomescape/common/exception/ExceptionController.java
@@ -1,0 +1,14 @@
+package roomescape.common.exception;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+@ControllerAdvice
+public class ExceptionController {
+    @ExceptionHandler(Exception.class)
+    public ResponseEntity<Void> handleRuntimeException(Exception e) {
+        e.printStackTrace();
+        return ResponseEntity.badRequest().build();
+    }
+}
