--- src/main/java/roomescape/member/LoginCheckResponse.java
@@ -1,9 +1,0 @@
-package roomescape.member;
-
-public record LoginCheckResponse(
-        String name
-) {
-    public LoginCheckResponse(Member member) {
-        this(member.getName());
-    }
-}
