--- src/main/java/roomescape/member/LoginRequest.java
@@ -1,7 +1,0 @@
-package roomescape.member;
-
-public record LoginRequest(
-        String password,
-        String email
-) {
-}
