--- src/main/java/roomescape/member/LoginResponse.java
@@ -1,6 +1,0 @@
-package roomescape.member;
-
-public record LoginResponse(
-        String accessToken
-) {
-}
