--- src/main/java/roomescape/member/MemberRequest.java
@@ -1,19 +1,0 @@
-package roomescape.member;
-
-public class MemberRequest {
-    private String name;
-    private String email;
-    private String password;
-
-    public String getName() {
-        return name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-}
