+++ src/main/java/roomescape/member/dto/request/LoginRequest.java
@@ -1,0 +1,7 @@
+package roomescape.member.dto.request;
+
+public record LoginRequest(
+        String password,
+        String email
+) {
+}
