+++ src/main/java/roomescape/member/dto/request/MemberRequest.java
@@ -1,0 +1,19 @@
+package roomescape.member.dto.request;
+
+public class MemberRequest {
+    private String name;
+    private String email;
+    private String password;
+
+    public String getName() {
+        return name;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+}
