+++ src/main/java/roomescape/member/dto/response/LoginCheckResponse.java
@@ -1,0 +1,11 @@
+package roomescape.member.dto.response;
+
+import roomescape.member.domain.Member;
+
+public record LoginCheckResponse(
+        String name
+) {
+    public LoginCheckResponse(Member member) {
+        this(member.getName());
+    }
+}
