+++ src/main/java/roomescape/member/dto/response/LoginResponse.java
@@ -1,0 +1,6 @@
+package roomescape.member.dto.response;
+
+public record LoginResponse(
+        String accessToken
+) {
+}
