--- src/main/java/roomescape/reservation/ReservationRequest.java
@@ -1,24 +1,0 @@
-package roomescape.reservation;
-
-public class ReservationRequest {
-    private String name;
-    private String date;
-    private Long theme;
-    private Long time;
-
-    public String getName() {
-        return name;
-    }
-
-    public String getDate() {
-        return date;
-    }
-
-    public Long getTheme() {
-        return theme;
-    }
-
-    public Long getTime() {
-        return time;
-    }
-}
