--- src/main/java/roomescape/reservation/ReservationResponse.java
@@ -1,37 +1,0 @@
-package roomescape.reservation;
-
-public class ReservationResponse {
-    private Long id;
-    private String name;
-    private String theme;
-    private String date;
-    private String time;
-
-    public ReservationResponse(Long id, String name, String theme, String date, String time) {
-        this.id = id;
-        this.name = name;
-        this.theme = theme;
-        this.date = date;
-        this.time = time;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getTheme() {
-        return theme;
-    }
-
-    public String getDate() {
-        return date;
-    }
-
-    public String getTime() {
-        return time;
-    }
-}
