+++ src/main/java/roomescape/reservation/dto/request/ReservationRequest.java
@@ -1,0 +1,24 @@
+package roomescape.reservation.dto.request;
+
+public class ReservationRequest {
+    private String name;
+    private String date;
+    private Long theme;
+    private Long time;
+
+    public String getName() {
+        return name;
+    }
+
+    public String getDate() {
+        return date;
+    }
+
+    public Long getTheme() {
+        return theme;
+    }
+
+    public Long getTime() {
+        return time;
+    }
+}
