+++ src/main/java/roomescape/theme/dao/ThemeDao.java
@@ -1,0 +1,42 @@
+package roomescape.theme.dao;
+
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.jdbc.support.KeyHolder;
+import org.springframework.stereotype.Repository;
+import roomescape.theme.domain.Theme;
+
+import java.util.List;
+
+@Repository
+public class ThemeDao {
+    private JdbcTemplate jdbcTemplate;
+
+    public ThemeDao(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    public List<Theme> findAll() {
+        return jdbcTemplate.query("SELECT * FROM theme where deleted = false", (rs, rowNum) -> new Theme(
+                rs.getLong("id"),
+                rs.getString("name"),
+                rs.getString("description")
+        ));
+    }
+
+    public Theme save(Theme theme) {
+        KeyHolder keyHolder = new GeneratedKeyHolder();
+        jdbcTemplate.update(connection -> {
+            var ps = connection.prepareStatement("INSERT INTO theme(name, description) VALUES (?, ?)", new String[]{"id"});
+            ps.setString(1, theme.getName());
+            ps.setString(2, theme.getDescription());
+            return ps;
+        }, keyHolder);
+
+        return new Theme(keyHolder.getKey().longValue(), theme.getName(), theme.getDescription());
+    }
+
+    public void deleteById(Long id) {
+        jdbcTemplate.update("UPDATE theme SET deleted = true WHERE id = ?", id);
+    }
+}
