+++ src/main/java/roomescape/time/service/TimeService.java
@@ -1,0 +1,47 @@
+package roomescape.time.service;
+
+import org.springframework.stereotype.Service;
+import roomescape.reservation.domain.Reservation;
+import roomescape.reservation.dao.ReservationDao;
+import roomescape.time.dao.TimeDao;
+import roomescape.time.domain.AvailableTime;
+import roomescape.time.domain.Time;
+
+import java.util.List;
+
+@Service
+public class TimeService {
+    private TimeDao timeDao;
+    private ReservationDao reservationDao;
+
+    public TimeService(TimeDao timeDao, ReservationDao reservationDao) {
+        this.timeDao = timeDao;
+        this.reservationDao = reservationDao;
+    }
+
+    public List<AvailableTime> getAvailableTime(String date, Long themeId) {
+        List<Reservation> reservations = reservationDao.findByDateAndThemeId(date, themeId);
+        List<Time> times = timeDao.findAll();
+
+        return times.stream()
+                .map(time -> new AvailableTime(
+                        time.getId(),
+                        time.getValue(),
+                        reservations.stream()
+                                .anyMatch(reservation -> reservation.getTime().getId().equals(time.getId()))
+                ))
+                .toList();
+    }
+
+    public List<Time> findAll() {
+        return timeDao.findAll();
+    }
+
+    public Time save(Time time) {
+        return timeDao.save(time);
+    }
+
+    public void deleteById(Long id) {
+        timeDao.deleteById(id);
+    }
+}
